+++ packages/natds-themes/config/shared/registerAssetInlineTransform.test.js
@@ -1,8 +1,13 @@
 /* eslint-disable mocha/max-top-level-suites */
 import fs from 'fs';
+import path from 'path';
+import svgo from 'svgo';
 import { registerAssetInlineTransform } from './registerAssetInlineTransform';
 
 jest.mock('fs');
+jest.mock('path');
+jest.mock('svgo');
+
 describe('registerAssetInlineTransform', () => {
   it('should return the transformer config', () => {
     const config = registerAssetInlineTransform();
@@ -20,8 +25,11 @@ describe('registerAssetInlineTransform', () => {
 describe('transformer', () => {
   beforeEach(() => {
     fs.readFileSync.mockReturnValue('<svg>natura-a-official</svg>');
+    path.join.mockReturnValue('a/path');
   });
+
   it('should return the svg content as string', () => {
+    svgo.optimize.mockReturnValue({});
     const config = registerAssetInlineTransform();
     const token = { original: { value: 'natura-a-official' } };
 
@@ -30,5 +38,6 @@ describe('transformer', () => {
     const expectedResult = '<svg>natura-a-official</svg>';
 
     expect(result).toEqual(expectedResult);
+    expect(fs.readFileSync).toHaveBeenCalledWith('a/path');
   });
 });